requirements.txt
alabaster==0.7.13
alembic==1.11.1
astroid==2.15.3
Babel==2.12.1
blinker==1.6.2
Cerberus==1.3.4
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.3
dill==0.3.6
distlib==0.3.6
docutils==0.18.1
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.3.2
greenlet==2.0.2
identify==2.5.22
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
iniconfig==2.0.0
install==1.3.5
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
lxml-stubs==0.4.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.2
psycopg2==2.9.6
Pygments==2.15.1
pylint==2.17.2
pylint-pytest==1.1.2
pytest==7.3.1
pytest-cov==4.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-flask==1.2.0
pytest-mock==3.10.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.15
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
types-docutils==0.20.0.1
typing_extensions==4.5.0
urllib3==2.0.3
virtualenv==20.22.0
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.15.0
